refactor(room): 폴더 명 변경
