+++ src/feature/room/index.tsx
@@ -1,1 +1,1 @@
-export * from "./content";
+export * from "./room";
